rename the utility functions
