+++ python/mxnet/io/utils.py
